+++ instructions/condition.h
@@ -6,11 +6,11 @@
 class Condition : public Instruction
 {
 public:
-    Condition(unsigned input_pointer, unsigned subprogram_pointer);
-    void operator()(std::vector<float>& memory, unsigned& subprogram_pointer, std::vector<unsigned>& instruction_pointers, unsigned& return_pointer) const override;
+    Condition(unsigned short input_pointer, unsigned short subprogram_pointer);
+    void operator()(std::vector<float>& memory, unsigned short& subprogram_pointer, std::vector<unsigned short>& instruction_pointers, unsigned short& return_pointer) const override;
 private:
-    unsigned m_input_pointer;
-    unsigned m_subprogram_pointer;
+    unsigned short m_input_pointer;
+    unsigned short m_subprogram_pointer;
 };
 
 #endif // CONDITION_H
