+++ instructions/loop.cpp
@@ -1,11 +1,11 @@
 #include "loop.h"
 
-Loop::Loop(unsigned subprogram_pointer) :
+Loop::Loop(unsigned short subprogram_pointer) :
     m_subprogram_pointer(subprogram_pointer)
 {
 }
 
-void Loop::operator()(std::vector<float>& memory, unsigned& subprogram_pointer, std::vector<unsigned>& instruction_pointers, unsigned& return_pointer) const
+void Loop::operator()(std::vector<float>& memory, unsigned short& subprogram_pointer, std::vector<unsigned short>& instruction_pointers, unsigned short& return_pointer) const
 {
     (void)memory;
     (void)return_pointer;
