+++ instructions/negate.h
@@ -6,11 +6,11 @@
 class Negate : public Instruction
 {
 public:
-    Negate(unsigned input_pointer, unsigned output_pointer);
-    void operator()(std::vector<float>& memory, unsigned& subprogram_pointer, std::vector<unsigned>& instruction_pointers, unsigned& return_pointer) const override;
+    Negate(unsigned short input_pointer, unsigned short output_pointer);
+    void operator()(std::vector<float>& memory, unsigned short& subprogram_pointer, std::vector<unsigned short>& instruction_pointers, unsigned short& return_pointer) const override;
 private:
-    unsigned m_input_pointer;
-    unsigned m_output_pointer;
+    unsigned short m_input_pointer;
+    unsigned short m_output_pointer;
 };
 
 #endif // NEGATE_H
