+++ instructions/return.cpp
@@ -1,6 +1,6 @@
 #include "return.h"
 
-void Return::operator()(std::vector<float>& memory, unsigned& subprogram_pointer, std::vector<unsigned>& instruction_pointers, unsigned& return_pointer) const
+void Return::operator()(std::vector<float>& memory, unsigned short& subprogram_pointer, std::vector<unsigned short>& instruction_pointers, unsigned short& return_pointer) const
 {
     (void)memory;
     (void)instruction_pointers;
